ig-prospector-mcp/pyproject.toml
[project]
name = "ig-prospector-mcp"
version = "0.1.0"
description = "MCP server to discover Instagram clothing-store prospects in Europe and generate personalized outreach"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.2.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.1",
    "anthropic>=0.39.0",
    "apify-client>=1.8.1",
    "firecrawl-py>=1.6.0",
]

[project.scripts]
ig-prospector-mcp = "ig_prospector.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ig_prospector"]
